requirements.txt
# FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# MongoDB & ODM
motor==3.6.0
beanie==1.28.0
pymongo>=4.9,<4.10

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.9.0

# Telegram Bot
python-telegram-bot==21.7

# Pydantic & Validation
pydantic==1.10.18
email-validator==2.2.0

# Utilities
python-dotenv==1.0.1

# Development & Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# Code Quality
ruff==0.7.1
black==24.10.0